requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
sentence_transformers
chainlit
langchain_community
fuzzywuzzy
chainlit
python-Levenshtein
